requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
redis[hiredis]>=5.0
presidio-analyzer>=2.2
presidio-anonymizer>=2.2
pydantic>=2.0
pydantic-settings>=2.0
python-multipart>=0.0.9
httpx>=0.27
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-dotenv>=1.0
python-docx>=1.1
pypdf>=4.0
openpyxl>=3.1
nats-py>=2.7.0
